refactor:download process
issue https://github.com/blblblblblblblblbl/Unsplash-API-app/issues/24
is done
lost feature with snackbar, now it works only with workmanager and notification, but it's better and faster.
Also download without workmanager didn't work in some cases(for exapmle close application).
