+++ myProject/app/src/main/java/com/blblblbl/myapplication/presentation/view/fragments/PhotoDetailedInfoFragment.kt
@@ -97,7 +97,7 @@ class PhotoDetailedInfoFragment : Fragment() {
                         )
                         when (snackbarResult) {
                             SnackbarResult.Dismissed -> TODO()
-                            SnackbarResult.ActionPerformed -> viewModel.openGallery()
+                            SnackbarResult.ActionPerformed -> TODO()//viewModel.openGallery()
                         }
                     }
                 }
